Implement Job Summaries
Using the Wiz API, we can pull the scan results back as JSON (the output
of the scan is not parse-friendly), and built a Job Summary that
indicates pass/fail, the policy violated, and its thresholds, along with
a link to the report on Wiz.

Unfortunately, this requires knowing your api-endpoint-url, for which we
have no sensible default. So we make that input optional and only take
this action if it's provided.

If there's any failure, we just warn and skip the summary. For example,
our CI currently fails this step due to a lack of scopes.