Listen for scanId on stdout and stderr
Previously, the CLI output everything to `stderr`. The latest version
has changed what it outputs, and the `scanId` output is in `stdout` now
causing our runs to fail when it doesn't find it -- without any changes
to our code, fun.

As far as I can tell, it's not doing things with much logic; I see info
stuff in `stderr` and error stuff in `stdout` in my testing, so we'll
listen on both streams (instead changing it to only `stdout`, which
would make a lot more sense).