fix: don't document error as a possible result
There's no reason to have this value; if the scan errors, that manifests
as the job failing. This should most often do what you want implicitly
(e.g. fail the workflow), and if you're trying to do something specific
in the case of an error, you can make that conditional on `job.status`.